Moved fixtures to the testfile
